+++ libmpv/src/main/cpp/jni_utils.cpp
@@ -43,6 +43,7 @@ void init_methods_cache(JNIEnv *env) {
         mpv_MPVLib_eventProperty_Sd = env->GetMethodID(mpv_MPVLib, "eventProperty", "(Ljava/lang/String;D)V"); // eventProperty(String, double)
         mpv_MPVLib_eventProperty_SS = env->GetMethodID(mpv_MPVLib, "eventProperty", "(Ljava/lang/String;Ljava/lang/String;)V"); // eventProperty(String, String)
         mpv_MPVLib_event = env->GetMethodID(mpv_MPVLib, "event", "(I)V"); // event(int)
+        mpv_MPVLib_endFile_II = env->GetMethodID(mpv_MPVLib, "endFile", "(II)V"); // endFile(int reason, int error)
         mpv_MPVLib_logMessage_SiS = env->GetMethodID(mpv_MPVLib, "logMessage", "(Ljava/lang/String;ILjava/lang/String;)V"); // logMessage(String, int, String)
     });
 }