+++ libmpv/src/main/cpp/jni_utils.h
@@ -22,4 +22,5 @@ UTIL_EXTERN jmethodID mpv_MPVLib_eventProperty_S,
         mpv_MPVLib_eventProperty_Sd,
         mpv_MPVLib_eventProperty_SS,
         mpv_MPVLib_event,
+        mpv_MPVLib_endFile_II,
         mpv_MPVLib_logMessage_SiS;